Unified destroy/remote.go and deploy/remote.go main logic to avoid duplicity
