+++ cmd/deploy/validate.go
@@ -14,31 +14,15 @@
 package deploy
 
 import (
-	"fmt"
-	"strings"
-
 	"github.com/okteto/okteto/pkg/env"
 )
 
 func validateAndSet(variables []string, setEnv func(key, value string) error) error {
-	envVars, err := parse(variables)
+	envVars, err := env.Parse(variables)
 	if err != nil {
 		return err
 	}
 	return setOptionVarsAsEnvs(envVars, setEnv)
-}
-
-func parse(variables []string) ([]env.Var, error) {
-	var result []env.Var
-	for _, v := range variables {
-		variableFormatParts := 2
-		kv := strings.SplitN(v, "=", variableFormatParts)
-		if len(kv) != variableFormatParts {
-			return nil, fmt.Errorf("invalid variable value '%s': must follow KEY=VALUE format", v)
-		}
-		result = append(result, env.Var{Name: kv[0], Value: kv[1]})
-	}
-	return result, nil
 }
 
 func setOptionVarsAsEnvs(variables []env.Var, setEnv func(key, value string) error) error {
