+++ cmd/destroy/remote.go
@@ -15,193 +15,63 @@ package destroy
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/base64"
 	"errors"
 	"fmt"
-	"math/big"
-	"net"
-	"os"
-	"path/filepath"
-	"regexp"
 	"strings"
-	"text/template"
 
-	"github.com/mitchellh/go-homedir"
-	"github.com/okteto/okteto/pkg/build"
 	buildCmd "github.com/okteto/okteto/pkg/cmd/build"
-	"github.com/okteto/okteto/pkg/config"
-	"github.com/okteto/okteto/pkg/constants"
 	"github.com/okteto/okteto/pkg/deployable"
+	"github.com/okteto/okteto/pkg/env"
 	oktetoErrors "github.com/okteto/okteto/pkg/errors"
-	"github.com/okteto/okteto/pkg/filesystem"
 	oktetoLog "github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/log/io"
 	"github.com/okteto/okteto/pkg/model"
 	"github.com/okteto/okteto/pkg/okteto"
 	"github.com/okteto/okteto/pkg/remote"
-	"github.com/okteto/okteto/pkg/types"
 	"github.com/spf13/afero"
-	"gopkg.in/yaml.v2"
 )
 
 const (
 	templateName           = "destroy-dockerfile"
 	dockerfileTemporalNane = "Dockerfile.destroy"
-	dockerfileTemplate     = `
-FROM {{ .OktetoCLIImage }} as okteto-cli
-
-FROM {{ .UserDestroyImage }} as deploy
-
-ENV PATH="${PATH}:/okteto/bin"
-COPY --from=okteto-cli /usr/local/bin/* /okteto/bin/
-
-ENV {{ .RemoteDeployEnvVar }} true
-
-ARG {{ .NamespaceArgName }}
-ARG {{ .ContextArgName }}
-ARG {{ .TokenArgName }}
-ARG {{ .ActionNameArgName }}
-ARG {{ .TlsCertBase64ArgName }}
-ARG {{ .InternalServerName }}
-ARG {{ .OktetoDeployable }}
-RUN mkdir -p /etc/ssl/certs/
-RUN echo "${{ .TlsCertBase64ArgName }}" | base64 -d > /etc/ssl/certs/okteto.crt
-
-COPY . /okteto/src
-WORKDIR /okteto/src
-
-ARG {{ .GitCommitArgName }}
-ARG {{ .GitBranchArgName }}
-ARG {{ .InvalidateCacheArgName }}
-
-RUN okteto registrytoken install --force --log-output=json
-
-RUN --mount=type=secret,id=known_hosts --mount=id=remote,type=ssh \
-  mkdir -p $HOME/.ssh && echo "UserKnownHostsFile=/run/secrets/known_hosts" >> $HOME/.ssh/config && \
-  /okteto/bin/okteto remote-run destroy --log-output=json --server-name="${{ .InternalServerName }}" {{ .CommandFlags }}
-`
+	destroyCMD             = "destroy"
 )
 
 // remoteRunner is the interface to run the destroy command remotely. The implementation is using
 // an image builder like BuildKit
 type remoteRunner interface {
-	Run(ctx context.Context, buildOptions *types.BuildOptions, ioCtrl *io.Controller) error
-}
-
-type dockerfileTemplateProperties struct {
-	OktetoCLIImage         string
-	UserDestroyImage       string
-	RemoteDeployEnvVar     string
-	ContextArgName         string
-	NamespaceArgName       string
-	TokenArgName           string
-	TlsCertBase64ArgName   string
-	InternalServerName     string
-	ActionNameArgName      string
-	GitCommitArgName       string
-	GitBranchArgName       string
-	InvalidateCacheArgName string
-	CommandFlags           string
-	OktetoDeployable       string
+	Run(ctx context.Context, params *remote.Params) error
 }
 
 type remoteDestroyCommand struct {
-	runner               remoteRunner
-	destroyImage         string
-	fs                   afero.Fs
-	workingDirectoryCtrl filesystem.WorkingDirectoryInterface
-	temporalCtrl         filesystem.TemporalDirectoryInterface
-	manifest             *model.Manifest
-	clusterMetadata      func(context.Context) (*types.ClusterMetadata, error)
+	runner   remoteRunner
+	manifest *model.Manifest
 
 	// ioCtrl is the controller for the output of the Build logs
 	ioCtrl *io.Controller
-
-	// sshAuthSockEnvvar is the default for SSH_AUTH_SOCK. Provided mostly for testing
-	sshAuthSockEnvvar string
-
-	// knownHostsPath  is the default known_hosts file path. Provided mostly for testing
-	knownHostsPath string
-
-	useInternalNetwork bool
 }
 
+// newRemoteDestroyer creates a new remote destroyer
 func newRemoteDestroyer(manifest *model.Manifest, ioCtrl *io.Controller) *remoteDestroyCommand {
 	fs := afero.NewOsFs()
-	runner := buildCmd.NewOktetoBuilder(
+	builder := buildCmd.NewOktetoBuilder(
 		&okteto.ContextStateless{
 			Store: okteto.GetContextStore(),
 		},
 		fs,
 	)
+	runner := remote.NewRunner(ioCtrl, builder)
 	if manifest.Destroy == nil {
 		manifest.Destroy = &model.DestroyInfo{}
 	}
 	return &remoteDestroyCommand{
-		runner:               runner,
-		destroyImage:         manifest.Destroy.Image,
-		fs:                   fs,
-		workingDirectoryCtrl: filesystem.NewOsWorkingDirectoryCtrl(),
-		temporalCtrl:         filesystem.NewTemporalDirectoryCtrl(fs),
-		manifest:             manifest,
-		clusterMetadata:      fetchClusterMetadata,
-		ioCtrl:               ioCtrl,
-		useInternalNetwork:   !buildCmd.IsDepotEnabled(),
+		runner:   runner,
+		manifest: manifest,
+		ioCtrl:   ioCtrl,
 	}
 }
 
 func (rd *remoteDestroyCommand) Destroy(ctx context.Context, opts *Options) error {
-	home, err := homedir.Dir()
-	if err != nil {
-		return err
-	}
-
-	sc, err := rd.clusterMetadata(ctx)
-	if err != nil {
-		return err
-	}
-
-	if rd.destroyImage == "" {
-		rd.destroyImage = sc.PipelineRunnerImage
-	}
-
-	cwd, err := rd.getOriginalCWD(opts.ManifestPathFlag)
-	if err != nil {
-		return err
-	}
-
-	tmpDir, err := rd.temporalCtrl.Create()
-	if err != nil {
-		return err
-	}
-
-	dockerfile, err := rd.createDockerfile(tmpDir, opts)
-	if err != nil {
-		return err
-	}
-
-	defer func() {
-		if err := rd.fs.Remove(dockerfile); err != nil {
-			oktetoLog.Infof("error removing dockerfile: %s", err)
-		}
-	}()
-
-	buildInfo := &build.Info{
-		Dockerfile: dockerfile,
-		Context:    rd.getContextPath(cwd, opts.ManifestPathFlag),
-	}
-
-	// undo modification of CWD for Build command
-	if err := rd.workingDirectoryCtrl.Change(cwd); err != nil {
-		return err
-	}
-
-	randomNumber, err := rand.Int(rand.Reader, big.NewInt(1000000))
-	if err != nil {
-		return err
-	}
-
 	if opts.Manifest == nil {
 		opts.Manifest = &model.Manifest{}
 	}
@@ -210,83 +80,41 @@ func (rd *remoteDestroyCommand) Destroy(ctx context.Context, opts *Options) erro
 		opts.Manifest.Destroy = &model.DestroyInfo{}
 	}
 
-	deployable := deployable.Entity{
+	baseImage := ""
+	if opts.Manifest.Destroy.Image != "" {
+		baseImage = opts.Manifest.Destroy.Image
+	}
+
+	dep := deployable.Entity{
 		Commands: opts.Manifest.Destroy.Commands,
 		// Added this for backward compatibility. Before the refactor we were having the env variables for the external
 		// resources in the environment, so including it to set the env vars in the remote-run
 		External: opts.Manifest.External,
 	}
 
-	b, err := yaml.Marshal(deployable)
+	commandFlags, err := getCommandFlags(opts)
 	if err != nil {
 		return err
 	}
 
-	buildOptions := buildCmd.OptsFromBuildInfoForRemoteDeploy(buildInfo, &types.BuildOptions{Path: cwd, OutputMode: buildCmd.DestroyOutputModeOnBuild})
-	buildOptions.Manifest = rd.manifest
-	buildOptions.BuildArgs = append(
-		buildOptions.BuildArgs,
-		fmt.Sprintf("%s=%s", model.OktetoContextEnvVar, okteto.GetContext().Name),
-		fmt.Sprintf("%s=%s", model.OktetoNamespaceEnvVar, okteto.GetContext().Namespace),
-		fmt.Sprintf("%s=%s", model.OktetoTokenEnvVar, okteto.GetContext().Token),
-		fmt.Sprintf("%s=%s", constants.OktetoTlsCertBase64EnvVar, base64.StdEncoding.EncodeToString(sc.Certificate)),
-		fmt.Sprintf("%s=%s", model.OktetoActionNameEnvVar, os.Getenv(model.OktetoActionNameEnvVar)),
-		fmt.Sprintf("%s=%s", constants.OktetoGitCommitEnvVar, os.Getenv(constants.OktetoGitCommitEnvVar)),
-		fmt.Sprintf("%s=%s", constants.OktetoGitBranchEnvVar, os.Getenv(constants.OktetoGitBranchEnvVar)),
-		fmt.Sprintf("%s=%d", constants.OktetoInvalidateCacheEnvVar, int(randomNumber.Int64())),
-		fmt.Sprintf("%s=%s", constants.OktetoDeployableEnvVar, base64.StdEncoding.EncodeToString(b)),
-	)
-
-	if rd.useInternalNetwork {
-		buildOptions.BuildArgs = append(
-			buildOptions.BuildArgs,
-			fmt.Sprintf("%s=%s", constants.OktetoInternalServerNameEnvVar, sc.ServerName),
-		)
-		if sc.ServerName != "" {
-			registryUrl := okteto.GetContext().Registry
-			subdomain := strings.TrimPrefix(registryUrl, "registry.")
-			ip, _, err := net.SplitHostPort(sc.ServerName)
-			if err != nil {
-				return fmt.Errorf("failed to parse server name network address: %w", err)
-			}
-			buildOptions.ExtraHosts = getExtraHosts(registryUrl, subdomain, ip, *sc)
-		}
-	}
-
-	sshSock := os.Getenv(rd.sshAuthSockEnvvar)
-	if sshSock == "" {
-		sshSock = os.Getenv("SSH_AUTH_SOCK")
-	}
-
-	if sshSock != "" {
-		if _, err := os.Stat(sshSock); err != nil {
-			oktetoLog.Debugf("Not mounting ssh agent. Error reading socket: %s", err.Error())
-		} else {
-			sshSession := types.BuildSshSession{Id: "remote", Target: sshSock}
-			buildOptions.SshSessions = append(buildOptions.SshSessions, sshSession)
-		}
-
-		// TODO: check if ~/.ssh/config exists and has UserKnownHostsFile defined
-		knownHostsPath := rd.knownHostsPath
-		if knownHostsPath == "" {
-			knownHostsPath = filepath.Join(home, ".ssh", "known_hosts")
-		}
-		if _, err := os.Stat(knownHostsPath); err != nil {
-			oktetoLog.Debugf("Not know_hosts file. Error reading file: %s", err.Error())
-		} else {
-			oktetoLog.Debugf("reading known hosts from %s", knownHostsPath)
-			buildOptions.Secrets = append(buildOptions.Secrets, fmt.Sprintf("id=known_hosts,src=%s", knownHostsPath))
-		}
-
-	} else {
-		oktetoLog.Debug("no ssh agent found. Not mouting ssh-agent for build")
+	runParams := remote.Params{
+		BaseImage:           baseImage,
+		ManifestPathFlag:    "",
+		TemplateName:        templateName,
+		CommandFlags:        commandFlags,
+		BuildEnvVars:        make(map[string]string),
+		DependenciesEnvVars: make(map[string]string),
+		DockerfileName:      dockerfileTemporalNane,
+		Deployable:          dep,
+		Manifest:            opts.Manifest,
+		Command:             destroyCMD,
 	}
 
 	// we need to call Run() method using a remote builder. This Builder will have
 	// the same behavior as the V1 builder but with a different output taking into
 	// account that we must not confuse the user with build messages since this logic is
 	// executed in the deploy command.
-	if err := rd.runner.Run(ctx, buildOptions, rd.ioCtrl); err != nil {
+	if err := rd.runner.Run(ctx, &runParams); err != nil {
 		var cmdErr buildCmd.OktetoCommandErr
 		if errors.As(err, &cmdErr) {
 			oktetoLog.SetStage(cmdErr.Stage)
@@ -309,64 +137,7 @@ func (rd *remoteDestroyCommand) Destroy(ctx context.Context, opts *Options) erro
 	return nil
 }
 
-func (rd *remoteDestroyCommand) createDockerfile(tempDir string, opts *Options) (string, error) {
-	cwd, err := rd.workingDirectoryCtrl.Get()
-	if err != nil {
-		return "", err
-	}
-
-	tmpl := template.Must(template.New(templateName).Parse(dockerfileTemplate))
-	dockerfileSyntax := dockerfileTemplateProperties{
-		OktetoCLIImage:         getOktetoCLIVersion(config.VersionString),
-		UserDestroyImage:       rd.destroyImage,
-		RemoteDeployEnvVar:     constants.OktetoDeployRemote,
-		ContextArgName:         model.OktetoContextEnvVar,
-		NamespaceArgName:       model.OktetoNamespaceEnvVar,
-		TokenArgName:           model.OktetoTokenEnvVar,
-		TlsCertBase64ArgName:   constants.OktetoTlsCertBase64EnvVar,
-		InternalServerName:     constants.OktetoInternalServerNameEnvVar,
-		ActionNameArgName:      model.OktetoActionNameEnvVar,
-		GitCommitArgName:       constants.OktetoGitCommitEnvVar,
-		GitBranchArgName:       constants.OktetoGitBranchEnvVar,
-		InvalidateCacheArgName: constants.OktetoInvalidateCacheEnvVar,
-		CommandFlags:           strings.Join(getCommandFlags(opts), " "),
-		OktetoDeployable:       constants.OktetoDeployableEnvVar,
-	}
-
-	dockerfile, err := rd.fs.Create(filepath.Join(tempDir, dockerfileTemporalNane))
-	if err != nil {
-		return "", err
-	}
-
-	if err = remote.CreateDockerignoreFileWithFilesystem(cwd, tempDir, filesystem.CleanManifestPath(opts.ManifestPathFlag), rd.fs); err != nil {
-		return "", err
-	}
-
-	if err := tmpl.Execute(dockerfile, dockerfileSyntax); err != nil {
-		return "", err
-	}
-	return dockerfile.Name(), nil
-
-}
-
-func getExtraHosts(registryURL, subdomain, ip string, metadata types.ClusterMetadata) []types.HostMap {
-	extraHosts := []types.HostMap{
-		{Hostname: registryURL, IP: ip},
-		{Hostname: fmt.Sprintf("kubernetes.%s", subdomain), IP: ip},
-	}
-
-	if metadata.BuildKitInternalIP != "" {
-		extraHosts = append(extraHosts, types.HostMap{Hostname: fmt.Sprintf("buildkit.%s", subdomain), IP: metadata.BuildKitInternalIP})
-	}
-
-	if metadata.PublicDomain != "" {
-		extraHosts = append(extraHosts, types.HostMap{Hostname: metadata.PublicDomain, IP: ip})
-	}
-
-	return extraHosts
-}
-
-func getCommandFlags(opts *Options) []string {
+func getCommandFlags(opts *Options) ([]string, error) {
 	commandFlags := []string{
 		fmt.Sprintf("--name \"%s\"", opts.Name),
 	}
@@ -375,82 +146,17 @@ func getCommandFlags(opts *Options) []string {
 		commandFlags = append(commandFlags, "--force-destroy")
 	}
 
-	return commandFlags
-}
-
-func getOktetoCLIVersion(versionString string) string {
-	var version string
-	if match, err := regexp.MatchString(`\d+\.\d+\.\d+`, versionString); match {
-		version = fmt.Sprintf(constants.OktetoCLIImageForRemoteTemplate, versionString)
-	} else {
+	if len(opts.Variables) > 0 {
+		var varsToAddForDeploy []string
+		variables, err := env.Parse(opts.Variables)
 		if err != nil {
-			oktetoLog.Infof("invalid okteto CLI version %s: %s. Using latest", versionString, err)
-		} else {
-			oktetoLog.Infof("invalid okteto CLI version %s. Using latest", versionString)
+			return nil, err
 		}
-		remoteOktetoImage := os.Getenv(constants.OktetoDeployRemoteImage)
-		if remoteOktetoImage != "" {
-			version = remoteOktetoImage
-		} else {
-			version = fmt.Sprintf(constants.OktetoCLIImageForRemoteTemplate, "latest")
+		for _, v := range variables {
+			varsToAddForDeploy = append(varsToAddForDeploy, fmt.Sprintf("--var %s=%q", v.Name, v.Value))
 		}
+		commandFlags = append(commandFlags, strings.Join(varsToAddForDeploy, " "))
 	}
 
-	return version
-}
-
-func fetchClusterMetadata(ctx context.Context) (*types.ClusterMetadata, error) {
-	cp := okteto.NewOktetoClientProvider()
-	c, err := cp.Provide()
-	if err != nil {
-		return nil, fmt.Errorf("failed to provide okteto client for fetching certs: %w", err)
-	}
-	uc := c.User()
-
-	metadata, err := uc.GetClusterMetadata(ctx, okteto.GetContext().Namespace)
-	if err != nil {
-		return nil, err
-	}
-
-	if metadata.Certificate == nil {
-		metadata.Certificate, err = uc.GetClusterCertificate(ctx, okteto.GetContext().Name, okteto.GetContext().Namespace)
-	}
-
-	return &metadata, err
-}
-
-func (rd *remoteDestroyCommand) getContextPath(cwd, manifestPath string) string {
-	if manifestPath == "" {
-		return cwd
-	}
-
-	path := manifestPath
-	if !filepath.IsAbs(manifestPath) {
-		path = filepath.Join(cwd, manifestPath)
-	}
-	fInfo, err := rd.fs.Stat(path)
-	if err != nil {
-		oktetoLog.Infof("error getting file info: %s", err)
-		return cwd
-
-	}
-	if fInfo.IsDir() {
-		return path
-	}
-
-	possibleCtx := filepath.Dir(path)
-	if strings.HasSuffix(possibleCtx, ".okteto") {
-		return filepath.Dir(possibleCtx)
-	}
-	return possibleCtx
-}
-
-// getOriginalCWD returns the original cwd
-func (rd *remoteDestroyCommand) getOriginalCWD(manifestPath string) (string, error) {
-	cwd, err := rd.workingDirectoryCtrl.Get()
-	if err != nil {
-		return "", err
-	}
-	manifestPathDir := filepath.Dir(filepath.Clean(fmt.Sprintf("/%s", manifestPath)))
-	return strings.TrimSuffix(cwd, manifestPathDir), nil
+	return commandFlags, nil
 }
