+++ cmd/destroy/remote_test.go
@@ -14,133 +14,10 @@
 package destroy
 
 import (
-	"context"
-	"fmt"
-	"os"
-	"path/filepath"
 	"testing"
 
-	"github.com/okteto/okteto/pkg/cmd/build"
-	"github.com/okteto/okteto/pkg/constants"
-	oktetoErrors "github.com/okteto/okteto/pkg/errors"
-	filesystem "github.com/okteto/okteto/pkg/filesystem/fake"
-	"github.com/okteto/okteto/pkg/log/io"
-	"github.com/okteto/okteto/pkg/types"
-	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
-
-type fakeRunner struct {
-	err           error
-	assertOptions func(o *types.BuildOptions)
-}
-
-func (f fakeRunner) Run(_ context.Context, opts *types.BuildOptions, _ *io.Controller) error {
-	if f.assertOptions != nil {
-		f.assertOptions(opts)
-	}
-	return f.err
-}
-
-func TestRemoteTest(t *testing.T) {
-	ctx := context.Background()
-	wdCtrl := filesystem.NewFakeWorkingDirectoryCtrl(filepath.Clean("/"))
-	fs := afero.NewMemMapFs()
-	tempCreator := filesystem.NewTemporalDirectoryCtrl(fs)
-
-	type config struct {
-		wd            filesystem.FakeWorkingDirectoryCtrlErrors
-		tempFsCreator error
-		options       *Options
-		builderErr    error
-	}
-	var tests = []struct {
-		config   config
-		expected error
-		name     string
-	}{
-		{
-			name: "OS can't access to the working directory",
-			config: config{
-				wd: filesystem.FakeWorkingDirectoryCtrlErrors{
-					Getter: assert.AnError,
-				},
-				options: &Options{},
-			},
-			expected: assert.AnError,
-		},
-		{
-			name: "OS can't create temporal directory",
-			config: config{
-				options:       &Options{},
-				tempFsCreator: assert.AnError,
-			},
-			expected: assert.AnError,
-		},
-		{
-			name: "OS can't change to the previous working directory",
-			config: config{
-				wd: filesystem.FakeWorkingDirectoryCtrlErrors{
-					Setter: assert.AnError,
-				},
-				options: &Options{},
-			},
-			expected: assert.AnError,
-		},
-		{
-			name: "build incorrect",
-			config: config{
-				options:    &Options{},
-				builderErr: assert.AnError,
-			},
-			expected: oktetoErrors.UserError{
-				E: fmt.Errorf("error during destroy of the development environment: %w", assert.AnError),
-			},
-		},
-		{
-			name: "build with command error",
-			config: config{
-				options: &Options{},
-				builderErr: build.OktetoCommandErr{
-					Stage: "test",
-					Err:   assert.AnError,
-				},
-			},
-			expected: oktetoErrors.UserError{
-				E: fmt.Errorf("error during development environment deployment: %w", assert.AnError),
-			},
-		},
-		{
-			name: "everything correct",
-			config: config{
-				options: &Options{},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			wdCtrl.SetErrors(tt.config.wd)
-			tempCreator.SetError(tt.config.tempFsCreator)
-			rdc := remoteDestroyCommand{
-				runner:               fakeRunner{err: tt.config.builderErr},
-				fs:                   fs,
-				workingDirectoryCtrl: wdCtrl,
-				temporalCtrl:         tempCreator,
-				destroyImage:         "",
-				clusterMetadata: func(ctx context.Context) (*types.ClusterMetadata, error) {
-					return &types.ClusterMetadata{
-						Certificate: []byte("cert"),
-						ServerName:  "1.2.3.4:443",
-					}, nil
-				},
-			}
-			err := rdc.Destroy(ctx, tt.config.options)
-			assert.Equal(t, tt.expected, err)
-		})
-	}
-}
 
 func TestGetCommandFlags(t *testing.T) {
 	type config struct {
@@ -179,380 +56,36 @@ func TestGetCommandFlags(t *testing.T) {
 			},
 			expected: []string{"--name \"test\"", "--force-destroy"},
 		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			flags := getCommandFlags(tt.config.opts)
-			assert.Equal(t, tt.expected, flags)
-		})
-	}
-}
-
-func TestCreateDockerfile(t *testing.T) {
-	wdCtrl := filesystem.NewFakeWorkingDirectoryCtrl(filepath.Clean("/"))
-	fs := afero.NewMemMapFs()
-	type config struct {
-		wd   filesystem.FakeWorkingDirectoryCtrlErrors
-		opts *Options
-	}
-	type expected struct {
-		err               error
-		dockerfileName    string
-		dockerFileContent string
-	}
-	var tests = []struct {
-		config   config
-		expected expected
-		name     string
-	}{
 		{
-			name: "OS can't access working directory",
-			config: config{
-				wd: filesystem.FakeWorkingDirectoryCtrlErrors{
-					Getter: assert.AnError,
-				},
-			},
-			expected: expected{
-				dockerfileName:    "",
-				err:               assert.AnError,
-				dockerFileContent: "",
-			},
-		},
-		{
-			name: "with dockerignore",
+			name: "variables set",
 			config: config{
 				opts: &Options{
 					Name: "test",
+					Variables: []string{
+						"a=b",
+						"c=d",
+					},
 				},
 			},
-			expected: expected{
-				dockerfileName: filepath.Clean("/test/Dockerfile.destroy"),
-				dockerFileContent: `
-FROM okteto/okteto:latest as okteto-cli
-
-FROM test-image as deploy
-
-ENV PATH="${PATH}:/okteto/bin"
-COPY --from=okteto-cli /usr/local/bin/* /okteto/bin/
-
-ENV OKTETO_DEPLOY_REMOTE true
-
-ARG OKTETO_NAMESPACE
-ARG OKTETO_CONTEXT
-ARG OKTETO_TOKEN
-ARG OKTETO_ACTION_NAME
-ARG OKTETO_TLS_CERT_BASE64
-ARG INTERNAL_SERVER_NAME
-ARG OKTETO_DEPLOYABLE
-RUN mkdir -p /etc/ssl/certs/
-RUN echo "$OKTETO_TLS_CERT_BASE64" | base64 -d > /etc/ssl/certs/okteto.crt
-
-COPY . /okteto/src
-WORKDIR /okteto/src
-
-ARG OKTETO_GIT_COMMIT
-ARG OKTETO_GIT_BRANCH
-ARG OKTETO_INVALIDATE_CACHE
-
-RUN okteto registrytoken install --force --log-output=json
-
-RUN --mount=type=secret,id=known_hosts --mount=id=remote,type=ssh \
-  mkdir -p $HOME/.ssh && echo "UserKnownHostsFile=/run/secrets/known_hosts" >> $HOME/.ssh/config && \
-  /okteto/bin/okteto remote-run destroy --log-output=json --server-name="$INTERNAL_SERVER_NAME" --name "test"
-`,
+			expected: []string{"--name \"test\"", "--var a=\"b\" --var c=\"d\""},
+		},
+		{
+			name: "multiword var value",
+			config: config{
+				opts: &Options{
+					Name:      "test",
+					Variables: []string{"test=multi word value"},
+				},
 			},
+			expected: []string{"--name \"test\"", "--var test=\"multi word value\""},
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			wdCtrl.SetErrors(tt.config.wd)
-			rdc := remoteDestroyCommand{
-				fs:                   fs,
-				destroyImage:         "test-image",
-				workingDirectoryCtrl: wdCtrl,
-			}
-			dockerfileName, err := rdc.createDockerfile("/test", tt.config.opts)
-			assert.ErrorIs(t, err, tt.expected.err)
-			assert.Equal(t, tt.expected.dockerfileName, dockerfileName)
-
-			if tt.expected.err == nil {
-				_, err = rdc.fs.Stat(filepath.Join("/", "test", dockerfileTemporalNane))
-				assert.NoError(t, err)
-				content, err := afero.ReadFile(rdc.fs, filepath.Join("/", "test", dockerfileTemporalNane))
-				assert.NoError(t, err)
-				assert.Equal(t, tt.expected.dockerFileContent, string(content))
-			}
-
+			flags, err := getCommandFlags(tt.config.opts)
+			assert.Equal(t, tt.expected, flags)
+			assert.NoError(t, err)
 		})
 	}
-}
-
-func Test_getOktetoCLIVersion(t *testing.T) {
-	var tests = []struct {
-		name                                 string
-		versionString, expected, cliImageEnv string
-	}{
-		{
-			name:          "no version string and no env return latest",
-			versionString: "",
-			expected:      "okteto/okteto:latest",
-		},
-		{
-			name:          "no version string return env value",
-			versionString: "",
-			cliImageEnv:   "okteto/remote:test",
-			expected:      "okteto/remote:test",
-		},
-		{
-			name:          "found version string",
-			versionString: "2.2.2",
-			expected:      "okteto/okteto:2.2.2",
-		},
-		{
-			name:          "found incorrect version string return latest ",
-			versionString: "2.a.2",
-			expected:      "okteto/okteto:latest",
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			if tt.cliImageEnv != "" {
-				t.Setenv(constants.OktetoDeployRemoteImage, tt.cliImageEnv)
-			}
-
-			version := getOktetoCLIVersion(tt.versionString)
-			require.Equal(t, version, tt.expected)
-		})
-	}
-}
-
-func TestRemoteDestroyWithSshAgent(t *testing.T) {
-	fs := afero.NewMemMapFs()
-	socket, err := os.CreateTemp("", "okteto-test-ssh-*")
-	require.NoError(t, err)
-	defer socket.Close()
-
-	knowHostFile, err := os.CreateTemp("", "okteto-test-know_hosts-*")
-	require.NoError(t, err)
-	defer socket.Close()
-
-	assertFn := func(o *types.BuildOptions) {
-		assert.Contains(t, o.SshSessions, types.BuildSshSession{Id: "remote", Target: socket.Name()})
-		assert.Contains(t, o.Secrets, fmt.Sprintf("id=known_hosts,src=%s", knowHostFile.Name()))
-	}
-
-	envvarName := fmt.Sprintf("TEST_SOCKET_%s", os.Getenv("RANDOM"))
-
-	t.Setenv(envvarName, socket.Name())
-	defer func() {
-		t.Logf("cleaning up %s envvar", envvarName)
-		os.Unsetenv(envvarName)
-	}()
-	rdc := remoteDestroyCommand{
-		sshAuthSockEnvvar:    envvarName,
-		knownHostsPath:       knowHostFile.Name(),
-		runner:               fakeRunner{assertOptions: assertFn},
-		fs:                   fs,
-		workingDirectoryCtrl: filesystem.NewFakeWorkingDirectoryCtrl(filepath.Clean("/")),
-		temporalCtrl:         filesystem.NewTemporalDirectoryCtrl(fs),
-		clusterMetadata: func(context.Context) (*types.ClusterMetadata, error) {
-			return &types.ClusterMetadata{}, nil
-		},
-	}
-
-	assert.NoError(t, rdc.Destroy(context.Background(), &Options{}))
-}
-
-func TestRemoteDestroyWithBadSshAgent(t *testing.T) {
-	fs := afero.NewMemMapFs()
-
-	assertFn := func(o *types.BuildOptions) {
-		assert.NotContains(t, o.SshSessions, types.BuildSshSession{Id: "remote", Target: "bad-socket"})
-		assert.NotContains(t, o.Secrets, fmt.Sprintf("id=known_hosts,src=%s", "inexistent-file"))
-	}
-
-	envvarName := fmt.Sprintf("TEST_SOCKET_%s", os.Getenv("RANDOM"))
-
-	t.Setenv(envvarName, "bad-socket")
-	defer func() {
-		t.Logf("cleaning up %s envvar", envvarName)
-		os.Unsetenv(envvarName)
-	}()
-
-	rdc := remoteDestroyCommand{
-		sshAuthSockEnvvar:    envvarName,
-		knownHostsPath:       "inexistent-file",
-		runner:               fakeRunner{assertOptions: assertFn},
-		fs:                   fs,
-		workingDirectoryCtrl: filesystem.NewFakeWorkingDirectoryCtrl(filepath.Clean("/")),
-		temporalCtrl:         filesystem.NewTemporalDirectoryCtrl(fs),
-		clusterMetadata: func(context.Context) (*types.ClusterMetadata, error) {
-			return &types.ClusterMetadata{}, nil
-		},
-	}
-
-	assert.NoError(t, rdc.Destroy(context.Background(), &Options{}))
-}
-
-func TestGetExtraHosts(t *testing.T) {
-	registryURL := "registry.test.dev.okteto.net"
-	subdomain := "test.dev.okteto.net"
-	ip := "1.2.3.4"
-
-	var tests = []struct {
-		name     string
-		expected []types.HostMap
-		metadata types.ClusterMetadata
-	}{
-		{
-			name:     "no metadata information",
-			metadata: types.ClusterMetadata{},
-			expected: []types.HostMap{
-				{Hostname: registryURL, IP: ip},
-				{Hostname: fmt.Sprintf("kubernetes.%s", subdomain), IP: ip},
-			},
-		},
-		{
-			name: "with buildkit internal ip",
-			metadata: types.ClusterMetadata{
-				BuildKitInternalIP: "4.3.2.1",
-			},
-			expected: []types.HostMap{
-				{Hostname: registryURL, IP: ip},
-				{Hostname: fmt.Sprintf("kubernetes.%s", subdomain), IP: ip},
-				{Hostname: fmt.Sprintf("buildkit.%s", subdomain), IP: "4.3.2.1"},
-			},
-		},
-		{
-			name: "with public domain",
-			metadata: types.ClusterMetadata{
-				PublicDomain: "publicdomain.dev.okteto.net",
-			},
-			expected: []types.HostMap{
-				{Hostname: registryURL, IP: ip},
-				{Hostname: fmt.Sprintf("kubernetes.%s", subdomain), IP: ip},
-				{Hostname: "publicdomain.dev.okteto.net", IP: ip},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			extraHosts := getExtraHosts(registryURL, subdomain, ip, tt.metadata)
-
-			assert.EqualValues(t, tt.expected, extraHosts)
-		})
-	}
-}
-
-func TestGetContextPath(t *testing.T) {
-	cwd := filepath.Clean("/path/to/current/directory")
-
-	rd := remoteDestroyCommand{
-		fs: afero.NewMemMapFs(),
-	}
-
-	t.Run("Manifest path is empty", func(t *testing.T) {
-		expected := cwd
-		result := rd.getContextPath(cwd, "")
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("Manifest path is a absolute path and directory", func(t *testing.T) {
-		manifestPath := filepath.Clean("/path/to/current/directory")
-		expected := manifestPath
-		rd.fs = afero.NewMemMapFs()
-		rd.fs.MkdirAll(manifestPath, 0755)
-		result := rd.getContextPath(cwd, manifestPath)
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("Manifest path is a file and absolute path", func(t *testing.T) {
-		manifestPath := filepath.Clean("/path/to/current/directory/file.yaml")
-		expected := filepath.Clean("/path/to/current/directory")
-		rd.fs = afero.NewMemMapFs()
-		rd.fs.MkdirAll(expected, 0755)
-		rd.fs.Create(manifestPath)
-		result := rd.getContextPath(cwd, manifestPath)
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("Manifest path is pointing to a file in the .okteto folder and absolute path", func(t *testing.T) {
-		manifestPath := filepath.Clean("/path/to/current/directory/.okteto/file.yaml")
-		expected := filepath.Clean("/path/to/current/directory")
-		rd.fs = afero.NewMemMapFs()
-		rd.fs.MkdirAll(expected, 0755)
-		rd.fs.Create(manifestPath)
-		result := rd.getContextPath(cwd, manifestPath)
-		assert.Equal(t, expected, result)
-	})
-
-	t.Run("Manifest path does not exist", func(t *testing.T) {
-		expected := cwd
-		result := rd.getContextPath(cwd, "nonexistent.yaml")
-		assert.Equal(t, expected, result)
-	})
-}
-
-func TestGetOriginalCWD(t *testing.T) {
-
-	t.Run("error getting the working directory", func(t *testing.T) {
-		wdCtrl := filesystem.NewFakeWorkingDirectoryCtrl(filepath.Clean("/"))
-		wdCtrl.SetErrors(filesystem.FakeWorkingDirectoryCtrlErrors{
-			Getter: assert.AnError,
-		})
-		deployCommand := &remoteDestroyCommand{
-			workingDirectoryCtrl: wdCtrl,
-		}
-
-		_, err := deployCommand.getOriginalCWD("")
-
-		require.Error(t, err)
-	})
-
-	t.Run("with empty manifest path", func(t *testing.T) {
-		wdCtrl := filesystem.NewFakeWorkingDirectoryCtrl(filepath.Clean("/tmp/test"))
-		deployCommand := &remoteDestroyCommand{
-			workingDirectoryCtrl: wdCtrl,
-		}
-
-		result, err := deployCommand.getOriginalCWD("")
-
-		expected := filepath.Clean("/tmp/test")
-		require.NoError(t, err)
-		require.Equal(t, expected, result)
-	})
-
-	t.Run("with manifest path to a dir", func(t *testing.T) {
-		wdCtrl := filesystem.NewFakeWorkingDirectoryCtrl(filepath.Clean("/tmp/test"))
-		deployCommand := &remoteDestroyCommand{
-			workingDirectoryCtrl: wdCtrl,
-		}
-
-		path := filepath.Join("test", ".okteto")
-		result, err := deployCommand.getOriginalCWD(path)
-
-		expected := filepath.Clean("/tmp")
-		require.NoError(t, err)
-		require.Equal(t, expected, result)
-	})
-
-	t.Run("with manifest path to a file", func(t *testing.T) {
-		wdCtrl := filesystem.NewFakeWorkingDirectoryCtrl(filepath.Clean("/tmp/test"))
-		deployCommand := &remoteDestroyCommand{
-			workingDirectoryCtrl: wdCtrl,
-		}
-
-		path := filepath.Join("test", "okteto.yml")
-		result, err := deployCommand.getOriginalCWD(path)
-
-		expected := filepath.Clean("/tmp")
-		require.NoError(t, err)
-		require.Equal(t, expected, result)
-	})
-
 }
